// LDC.Atlas.WebApp/Client/app/shared/entities/filter-gem.entity.ts
import { FilterValueGenerateEndMonth } from './filter-value-generate-end-month.entity';

export class FilterGem {
    departmentCode: FilterValueGenerateEndMonth[];
    charterCode: FilterValueGenerateEndMonth[];
    costType: FilterValueGenerateEndMonth[];
    associatedClient: FilterValueGenerateEndMonth[];
    dealCurrency: FilterValueGenerateEndMonth[];
    settlementCurrency: FilterValueGenerateEndMonth[];
    dealNumber: FilterValueGenerateEndMonth[];
}